// packages/react-docs/src/themes/theme-base/types.ts
export type AliasColorTokens = {
  // https://www.figma.com/file/KB9oUjMKen2cKnyPG7RgdS/Design-tokens-superset?node-id=1963%3A17486
  neutralForeground1: string;
  neutralForeground2: string;
  neutralForeground2Hover: string;
  neutralForeground2Pressed: string;
  neutralForeground2Selected: string;
  brandForeground2Hover: string;
  brandForeground2Pressed: string;
  brandForeground2Selected: string;
  neutralForeground3: string;
  neutralForeground3Hover: string;
  neutralForeground3Pressed: string;
  neutralForeground3Selected: string;
  brandForeground3Hover: string;
  brandForeground3Pressed: string;
  brandForeground3Selected: string;
  neutralForeground4: string;
  neutralForegroundDisabled: string;
  brandForeground: string;
  brandForegroundHover: string;
  brandForegroundPressed: string;
  brandForegroundSelected: string;
  neutralForegroundInverted: string;
  neutralForegroundInvertedAccessible: string;
  neutralBackground1: string;
  neutralBackground1Hover: string;
  neutralBackground1Pressed: string;
  neutralBackground1Selected: string;
  neutralBackground2: string;
  neutralBackground2Hover: string;
  neutralBackground2Pressed: string;
  neutralBackground2Selected: string;
  neutralBackground3: string;
  neutralBackground3Hover: string;
  neutralBackground3Pressed: string;
  neutralBackground3Selected: string;
  neutralBackground4: string;
  neutralBackground4Hover: string;
  neutralBackground4Pressed: string;
  neutralBackground4Selected: string;
  neutralBackground5: string;
  neutralBackground5Hover: string;
  neutralBackground5Pressed: string;
  neutralBackground5Selected: string;
  neutralBackground6: string;
  neutralBackgroundDisabled: string;
  neutralStrokeAccessible: string;
  neutralStrokeAccessibleHover: string;
  neutralStrokeAccessiblePressed: string;
  neutralStrokeAccessibleSelected: string;
  neutralStroke1: string;
  neutralStroke1Hover: string;
  neutralStroke1Pressed: string;
  neutralStroke1Selected: string;
  neutralStroke2: string;
  neutralStroke3: string;
  neutralStrokeDisabled: string;
  strokeAccessible: string;
  strokeAccessibleInteractive: string;
  strokeAccessibleDisabled: string;
  neutralShadowAmbient: string;
  neutralShadowKey: string;
  neutralShadowAmbientLighter: string;
  neutralShadowKeyLighter: string;
  neutralShadowAmbientDarker: string;
  neutralShadowKeyDarker: string;
};